move secret creation to always be at the root level, and simplify model creation so we're always directly manipulating a model
