+++ ui/app/templates/components/secret-list-header.hbs
@@ -32,8 +32,8 @@
           <div class="control">
             {{#secret-link
               mode="create"
-              secret=(or baseKey.id '')
-              queryParams=(query-params initialKey='')
+              secret=''
+              queryParams=(query-params initialKey=baseKey.id)
               class="button has-icon-right is-ghost is-compact"
               data-test-secret-create=true
             }}
